+++ CHANGELOG.rst
@@ -1,9 +1,11 @@
 Changelog
 =========
 
-Current (2025-10-16)
+Current (2026-08-09)
 --------------------
 
+* Implemented some CLI entry points to inspect software versions used to produce files (both binary and decoded)
+* Implemented the new binary versions and made it more robust, to hopefully not be as fragile in the future and be faster to update
 * Added new command line tool to convert data file formats
 * Added a new experimental option for processing a record batch into root without going through a pandas data frame