+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Current (2026-08-09)
 * Implemented the new binary versions and made it more robust, to hopefully not be as fragile in the future and be faster to update
 * Added new command line tool to convert data file formats
 * Added a new experimental option for processing a record batch into root without going through a pandas data frame
+* Added the new hit time unwrap logic
 
 
 0.1.2 (2025-10-16)