+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Current (2026-09-09)
 * Added the new hit time unwrap logic
 * Added a guard to not accidentally try processing the trigger file as data
 * Fixed metadata parsing for root files
+* Updated plot_hit_rate and plot_hit_rate entrypoint to be compatible with new unified format
 
 
 0.1.2 (2025-10-16)