+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Current (2026-09-09)
 * Fixed metadata parsing for root files
 * Updated plot_hit_rate and plot_hit_rate entrypoint to be compatible with new unified format
 * Updated plot_channel_hit_rate and plot_channel_hit_rate entrypoint to be compatible with new unified format
+* Updated the check time ordering and reorder data functionality to be compatible with the new unified format
 
 
 0.1.2 (2025-10-16)